requirements.txt
discord.py==2.5.1
yt-dlp==2023.9.28
google-api-python-client==2.86.0
python-dotenv==0.21.0
PyNaCl==1.4.0
